Refactor of inference services
