+++ edsl/inference_services/registry.py
@@ -1,45 +1,8 @@
-from edsl.inference_services.InferenceServicesCollection import (
-    InferenceServicesCollection,
-)
+from .inference_services_collection import InferenceServicesCollection
 
-from edsl.inference_services.OpenAIService import OpenAIService
-from edsl.inference_services.AnthropicService import AnthropicService
-from edsl.inference_services.DeepInfraService import DeepInfraService
-from edsl.inference_services.GoogleService import GoogleService
-from edsl.inference_services.GroqService import GroqService
-from edsl.inference_services.AwsBedrock import AwsBedrockService
-from edsl.inference_services.AzureAI import AzureAIService
-from edsl.inference_services.OllamaService import OllamaService
-from edsl.inference_services.TestService import TestService
-from edsl.inference_services.TogetherAIService import TogetherAIService
-from edsl.inference_services.PerplexityService import PerplexityService
-from edsl.inference_services.DeepSeekService import DeepSeekService
-from edsl.inference_services.XAIService import XAIService
+from .services import *
+from .services import __all__
 
-try:
-    from edsl.inference_services.MistralAIService import MistralAIService
-
-    mistral_available = True
-except Exception as e:
-    mistral_available = False
-
-services = [
-    OpenAIService,
-    AnthropicService,
-    DeepInfraService,
-    GoogleService,
-    GroqService,
-    AwsBedrockService,
-    AzureAIService,
-    OllamaService,
-    TestService,
-    TogetherAIService,
-    PerplexityService,
-    DeepSeekService,
-    XAIService,
-]
-
-if mistral_available:
-    services.append(MistralAIService)
+services = [globals()[service_name] for service_name in __all__ if service_name in globals()]
 
 default = InferenceServicesCollection(services)
