+++ edsl/inference_services/services/aws_bedrock.py
@@ -3,7 +3,7 @@ from typing import Any, List, Optional
 import re
 import boto3
 from botocore.exceptions import ClientError
-from edsl.inference_services.InferenceServiceABC import InferenceServiceABC
+from .inference_service_abc import InferenceServiceABC
 from ..language_models import LanguageModel
 import json
 from edsl.utilities.utilities import fix_partial_correct_response
