+++ edsl/inference_services/services/azure_ai.py
@@ -2,7 +2,7 @@ import os
 from typing import Any, Optional, List
 import re
 from openai import AsyncAzureOpenAI
-from edsl.inference_services.InferenceServiceABC import InferenceServiceABC
+from .inference_service_abc import InferenceServiceABC
 from ..language_models import LanguageModel
 
 from azure.ai.inference.aio import ChatCompletionsClient
