+++ edsl/language_models/__init__.py
@@ -2,4 +2,6 @@ from .language_model import LanguageModel
 from .model import Model
 from .model_list import ModelList
 
-__all__ = ["Model", "ModelList"]
+from .exceptions import LanguageModelBadResponseError
+
+__all__ = ["Model", "ModelList", "LanguageModelBadResponseError"]
